Update from original

+++ Library/IfActions.cpp
@@ -14,8 +14,8 @@
 #include "pcre/pcre.h"
 
 IfActions::IfActions() :
-	m_AboveValue(0.0f),
-	m_BelowValue(0.0f),
+	m_AboveValue(0.0),
+	m_BelowValue(0.0),
 	m_EqualValue(0),
 	m_AboveAction(),
 	m_BelowAction(),
@@ -37,13 +37,13 @@ IfActions::~IfActions()
 void IfActions::ReadOptions(ConfigParser& parser, const WCHAR* section)
 {
 	m_AboveAction = parser.ReadString(section, L"IfAboveAction", L"", false);
-	m_AboveValue = parser.ReadFloat(section, L"IfAboveValue", 0.0f);
+	m_AboveValue = parser.ReadFloat(section, L"IfAboveValue", 0.0);
 
 	m_BelowAction = parser.ReadString(section, L"IfBelowAction", L"", false);
-	m_BelowValue = parser.ReadFloat(section, L"IfBelowValue", 0.0f);
+	m_BelowValue = parser.ReadFloat(section, L"IfBelowValue", 0.0);
 
 	m_EqualAction = parser.ReadString(section, L"IfEqualAction", L"", false);
-	m_EqualValue = (int64_t)parser.ReadFloat(section, L"IfEqualValue", 0.0f);
+	m_EqualValue = (int64_t)parser.ReadFloat(section, L"IfEqualValue", 0.0);
 }
 
 void IfActions::ReadConditionOptions(ConfigParser& parser, const WCHAR* section)
@@ -200,7 +200,7 @@ void IfActions::DoIfActions(Measure& measure, double value)
 		++i;
 		if (!item.value.empty() && (!item.tAction.empty() || !item.fAction.empty()))
 		{
-			double result = 0.0f;
+			double result = 0.0;
 			const WCHAR* errMsg = MathParser::Parse(
 				item.value.c_str(), &result, measure.GetCurrentMeasureValue, &measure);
 			if (errMsg != nullptr)
@@ -222,7 +222,7 @@ void IfActions::DoIfActions(Measure& measure, double value)
 			{
 				item.parseError = false;
 
-				if (result == 1.0f)			// "True"
+				if (result == 1.0)			// "True"
 				{
 					item.fCommitted = false;
 
@@ -232,7 +232,7 @@ void IfActions::DoIfActions(Measure& measure, double value)
 						GetRainmeter().ExecuteCommand(item.tAction.c_str(), measure.GetSkin());
 					}
 				}
-				else if (result == 0.0f)	// "False"
+				else if (result == 0.0)	// "False"
 				{
 					item.tCommitted = false;
 
@@ -284,7 +284,7 @@ void IfActions::DoIfActions(Measure& measure, double value)
 				item.parseError = false;
 
 				const WCHAR* str = measure.GetStringValue();
-				int strLen = str ? wcslen(str) : 0;
+				int strLen = str ? (int)wcslen(str) : 0;
 				int ovector[300];
 				int rc = pcre16_exec(
 					re,
